backend/requirements.txt
requests
beautifulsoup4
lxml
dnspython
reportlab
python-dotenv
Flask
Flask-CORS
groq
redis
